+++ R/do_DotPlot.R
@@ -195,7 +195,9 @@ do_DotPlot <- function(sample,
     
 
     
-    data <- sample@assays[[assay]][[slot]][features, , drop = FALSE] %>% 
+    data <- Seurat::GetAssayData(object = sample,
+                                 assay = assay,
+                                 slot = slot)[features, , drop = FALSE] %>% 
             as.data.frame() %>% 
             tibble::rownames_to_column(var = "Gene") %>% 
             tidyr::pivot_longer(cols = -"Gene",